use size_t in most problematic places.

+++ src/unabto/unabto_stream_environment.c
@@ -13,10 +13,10 @@
 NABTO_THREAD_LOCAL_STORAGE unabto_stream stream__[NABTO_MEMORY_STREAM_MAX_STREAMS];  /**< a pool of streams */
 
 NABTO_THREAD_LOCAL_STORAGE bool stream_segment_pool[NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE];
-NABTO_THREAD_LOCAL_STORAGE uint8_t stream_buffer_data[NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE * NABTO_MEMORY_STREAM_SEGMENT_SIZE ];
+NABTO_THREAD_LOCAL_STORAGE uint8_t stream_buffer_data[(size_t)(NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE) * (size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE) ];
 
-NABTO_THREAD_LOCAL_STORAGE x_buffer x_buffers[NABTO_MEMORY_STREAM_MAX_STREAMS * NABTO_MEMORY_STREAM_SEND_WINDOW_SIZE];
-NABTO_THREAD_LOCAL_STORAGE r_buffer r_buffers[NABTO_MEMORY_STREAM_MAX_STREAMS * NABTO_MEMORY_STREAM_RECEIVE_WINDOW_SIZE];
+NABTO_THREAD_LOCAL_STORAGE x_buffer x_buffers[(size_t)(NABTO_MEMORY_STREAM_MAX_STREAMS) * (size_t)(NABTO_MEMORY_STREAM_SEND_WINDOW_SIZE)];
+NABTO_THREAD_LOCAL_STORAGE r_buffer r_buffers[(size_t)(NABTO_MEMORY_STREAM_MAX_STREAMS) * (size_t)(NABTO_MEMORY_STREAM_RECEIVE_WINDOW_SIZE)];
 #endif
 
 NABTO_THREAD_LOCAL_STORAGE int segment_pool_used_for_send_windows = 0;
@@ -86,8 +86,8 @@ bool unabto_stream_init_buffers(struct nabto_stream_s* stream)
     uint16_t recvWinSize = tcb->cfg.recvWinSize;
     uint16_t recvSegmentSize = tcb->cfg.recvPacketSize;
 
-    tcb->xmit = &x_buffers[unabto_stream_index(stream) * NABTO_MEMORY_STREAM_SEND_WINDOW_SIZE];
-    tcb->recv = &r_buffers[unabto_stream_index(stream) * NABTO_MEMORY_STREAM_RECEIVE_WINDOW_SIZE];
+    tcb->xmit = &x_buffers[(size_t)(unabto_stream_index(stream)) * (size_t)(NABTO_MEMORY_STREAM_SEND_WINDOW_SIZE)];
+    tcb->recv = &r_buffers[(size_t)(unabto_stream_index(stream)) * (size_t)(NABTO_MEMORY_STREAM_RECEIVE_WINDOW_SIZE)];
 
     for (i = 0; i < NABTO_MEMORY_STREAM_SEND_WINDOW_SIZE; i++)
     {
@@ -140,7 +140,7 @@ uint8_t* unabto_stream_alloc_send_segment(size_t required)
         if (!stream_segment_pool[i]) {
             stream_segment_pool[i] = true;
             segment_pool_used_for_send_windows++;
-            return stream_buffer_data + (i * NABTO_MEMORY_STREAM_SEGMENT_SIZE);
+            return stream_buffer_data + (i * (size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE));
         }
     }
     NABTO_LOG_TRACE(("Cannot allocate send segment, segment pool is empty"));
@@ -157,11 +157,11 @@ void unabto_stream_free_send_segment(uint8_t* buffer)
         NABTO_LOG_FATAL(("invalid free pointer is before pool"));
     }
 
-    if (buffer > (stream_buffer_data + (NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE * NABTO_MEMORY_STREAM_SEGMENT_SIZE))) {
+    if (buffer > (stream_buffer_data + ((size_t)(NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE) * (size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE)))) {
         NABTO_LOG_FATAL(("invalid free pointer is outside of pool area. buffer: %p, stream_buffer_data: %p, segment_pool_size: %d, segment_size: %d", buffer, stream_buffer_data, NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE, NABTO_MEMORY_STREAM_SEGMENT_SIZE));
     }
 
-    idx = (buffer - stream_buffer_data)/NABTO_MEMORY_STREAM_SEGMENT_SIZE;
+    idx = (buffer - stream_buffer_data)/(size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE);
 
     if (!stream_segment_pool[idx]) {
         NABTO_LOG_ERROR(("invalid free segment is not allocated"));
@@ -200,7 +200,7 @@ uint8_t* unabto_stream_alloc_recv_segment(size_t required)
         if (!stream_segment_pool[i]) {
             stream_segment_pool[i] = true;
             segment_pool_used_for_receive_windows++;
-            return stream_buffer_data + (i * NABTO_MEMORY_STREAM_SEGMENT_SIZE);
+            return stream_buffer_data + (i * (size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE));
         }
     }
     return NULL;
@@ -216,11 +216,11 @@ void unabto_stream_free_recv_segment(uint8_t* buffer)
         NABTO_LOG_FATAL(("invalid free pointer is before pool"));
     }
 
-    if (buffer > (stream_buffer_data + (NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE * NABTO_MEMORY_STREAM_SEGMENT_SIZE))) {
+    if (buffer > (stream_buffer_data + ((size_t)(NABTO_MEMORY_STREAM_SEGMENT_POOL_SIZE) * (size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE)))) {
         NABTO_LOG_FATAL(("invalid free pointer is outside of pool area"));
     }
 
-    idx = (buffer - stream_buffer_data)/NABTO_MEMORY_STREAM_SEGMENT_SIZE;
+    idx = (buffer - stream_buffer_data)/(size_t)(NABTO_MEMORY_STREAM_SEGMENT_SIZE);
 
     if (!stream_segment_pool[idx]) {
         NABTO_LOG_ERROR(("invalid free segment is not allocated"));
